发布商品 增加SKU 介绍图

+++ toucan-shopping-cloud/services/toucan-shopping-product/src/main/java/com/toucan/shopping/cloud/product/controller/ShopProductApproveController.java
@@ -784,7 +784,7 @@ public class ShopProductApproveController {
                             {
                                 for(ShopProductApproveDescriptionImgVO shopProductApproveDescriptionImgVO:shopProductApproveDescriptionImgVOS) {
                                     if (shopProductApproveDescriptionImgVO.getType().intValue() == 2) {
-                                        if(shopProductApproveSkuVO.getId().intValue()==shopProductApproveDescriptionImgVO.getProductSkuId().intValue())
+                                        if(shopProductApproveSkuVO.getId().longValue()==shopProductApproveDescriptionImgVO.getProductSkuId().longValue())
                                         {
                                             shopProductApproveSkuVO.setDescriptionImgFilePath(shopProductApproveDescriptionImgVO.getFilePath());
                                         }
@@ -911,7 +911,7 @@ public class ShopProductApproveController {
                             {
                                 for(ShopProductApproveDescriptionImgVO shopProductApproveDescriptionImgVO:productApproveDescriptionImgVOS) {
                                     if (shopProductApproveDescriptionImgVO.getType().intValue() == 2) {
-                                        if(shopProductApproveSkuVO.getId().intValue()==shopProductApproveDescriptionImgVO.getProductSkuId().intValue())
+                                        if(shopProductApproveSkuVO.getId().longValue()==shopProductApproveDescriptionImgVO.getProductSkuId().longValue())
                                         {
                                             shopProductApproveSkuVO.setDescriptionImgFilePath(shopProductApproveDescriptionImgVO.getFilePath());
                                         }
